Fix INFO and BGREWRITEAOF
This fixes the parsing of the response of the INFO and BGREWRITEAOF
commands on OS X with the following Redis version:

     Redis server
     v=2.8.11
     sha=00000000:0
     malloc=libc
     bits=64
     build=cd43e547b41f72f

In order to start the `redis-server` with the configuration file
redis.conf I removed everything but the password. The rest will
default to the settings that's used by the version of redis-server
that's installed. This was required as my version of redis-server
wouldn't start with the configuration file as it were.